+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All notable changes to this project will be documented in this file.
 - IT24610824 `src/app/api/pickup-locations/route.js` - Get Pickup Locations API, this helps frontend dropdown 
 - IT24610824 - `src/app/api/users/me/route.js` - Implemented users CRUD
 - IT24610824 - `src/app/api/admin/pickup-locations/[id]/route.js` - Implemented del and put fucntions on pickup.
+- IT24610824 - `src/app/api/admin/pickup-locations/route.js` - Added create function to pickup location.
+
 ---
 
 ## [0.1.0] – Initial Project Setup